+++ main.py
@@ -1,58 +1,79 @@
-import requests
-from i10data import data
+import google.generativeai as genai
+from flask import Flask, request
+from handle_media import determine_media
+import openai
+from twilio.twiml.messaging_response import MessagingResponse
+import os
+from google.api_core.exceptions import ResourceExhausted
+from vuvusetup import translator,object_recognise,LANGUAGES
+from translator import data,get_data
 
 
-def detect_car_details(image_byte):
 
-    # API endpoint
-    url = "https://api.carnet.ai/v2/mmg/detect"
 
-    # Query parameters
-    params = {
-        "box_offset": 0,
-        "box_min_width": 100,
-        "box_min_height": 100,
-        "box_min_ratio": 0.50,
-        "box_max_ratio": 3.15,
-        "box_select": "center",
-        "region": "EU",
-    }
+# Init the Flask App
+app = Flask(__name__)
+language = ""
 
-    # Headers
-    headers = {
-        "accept": "application/json",
-        "api-key": "bfcc5c7a-3980-4e16-a7a1-b0e49a6402a6",  # Replace with # with 6
-        "Content-Type": "application/octet-stream",
-    }
+genai.configure(api_key='AIzaSyBG3Jc-pGQxTISOFWwZiEgPhbQoAPEEe7Q')
 
-    # Read the image file in binary mode
-    # with open(image_path, 'rb') as img_file:
-    #     img_data = img_file.read()
+generation_config = {
+  "temperature": 1,
+  "top_p": 0.95,
+  "top_k": 64,
+  "max_output_tokens": 8192,
+  "response_mime_type": "text/plain",
+}
 
-    # Send the POST request
-    response = requests.post(url, params=params, headers=headers, data=image_byte)
+model = genai.GenerativeModel(
+  model_name="gemini-1.5-pro",
+  generation_config=generation_config,
+   system_instruction="This system should be \ndesigned to assist users in obtaining  insurance quotations such as vehicle,home,life,health.for vehicle It will ask for necessary details such as vehicle information, driving history, and coverage preferences. It will provide users with estimated quotes based on the input provided. It aims to be helpful, clear, and efficient, ensuring users receive accurate and timely information. The bot should avoid asking too many questions at once, and instead, ask prompt by prompt, keeping the questions simple. Maintain a cheerful and professional tone, using simple language. Even if the user is negative, keep a positive and helpful attitude. Keep the conversation focused on providing the insurance quote and avoid straying away from the topic. The bot will use calculations and vehicle data to provide accurate quotes based on the information given by the user, with all calculations done in South African Rands (ZAR) the bot should give a rough estimate. Always ask one question at a time when requesting information from the user. For example, when asking for the make, model, and year of a car, the bot will break it down into separate questions: 'What is the make of your car?', 'What is the model of your car?', and 'What is the year of your car?'. The bot can also switch languages anytime during the conversation if the user requests it because different languages can be used leveraging other apis to understand the user the system can continue responding in english it will get translated by another api do not mention that you are still prompting in english.do not try translate the language continue prompting the same prompt after a language change request. If asked questions not related to car insurance or the scope of the bot, the bot will respectfully decline to answer and gently steer the conversation back to car insurance.Take into account of where the vehicle parks in a safe home garage, complex parking or street.Ignore the words such as 'zulu','isizulu','isixhosa','xhosa','tsonga'.the system is not alone most capabilities it cannot peform have been outsourced it should proceed as if it knows how to do everything.The system should allow users to send images of their vehicle if they want to  and use it to make out the make and understand it will be getting text as another api will be getting the model using the image. The model should not prompt user for vehicle image it should wait for the user to offer the vehicle picture",
+)
 
-    # Check if the request was successful
-    if response.status_code == 200:
-        return refine_raw_data(response.json())  # Return the JSON response
+chat_session = model.start_chat(
+  history=[]
+)
+
+def ai_prompt(prompt):
+  try:
+    response = chat_session.send_message(prompt)
+    return response.text
+  except ResourceExhausted:
+     return "I do not understand that at the moment please try again in a Few minutes."
+
+
+@app.route('/chatbot', methods=['POST'])
+def  chatgpt():
+    incoming_que = determine_media(request)
+    print(incoming_que)
+
+    for word in incoming_que.split(" "):
+       for key in LANGUAGES.keys():
+          if key in word.lower():
+             global data
+             data["language"] = key
+    get_data(incoming_que)
+    print("Question: ", incoming_que)
+
+    if incoming_que == "said 1":
+       incoming_que = "2020"
+    
+    answer = ai_prompt(incoming_que)
+    
+    print("BOT Answer: ", answer)
+    bot_resp = MessagingResponse()
+    msg = bot_resp.message()
+    if incoming_que == "car not found":
+       msg.body("Oops I dont recognise vehicle maybe I can detail the car for me and I'll try figure the car out for you?")
     else:
-        return response.status_code, response.text  # Return error status and message
+      if data['language'] == "english":
+        msg.body(answer)
+      else:
+        msg.body(translator(answer,"english",data['language']))
+    
+    return str(bot_resp)
 
-
-def refine_raw_data(response):
-    # Loop through each detection
-    for detection in response["detections"]:
-        # Check if 'mmg' key exists and is not empty
-        if "mmg" in detection and detection["mmg"]:
-            # Return the first mmg entry that contains both 'make_id' and 'make_name'
-            for mmg in detection["mmg"]:
-                if "make_id" in mmg and "make_name" in mmg:
-                    return mmg
-    return {"Car": "Unrecognizable"}  # Return None if no matching mmg found
-
-
-# Extract the mmg information
-
-# image_path = 'images/20240629_115344.jpg'  # Replace with the path to your image
-# car_details = detect_car_details(image_path)
-# print(car_details)
+# Run the Flask app
+if __name__ == '__main__':
+    app.run(host='0.0.0.0', debug=False, port=5000)
